Address PR feedback for Apply Edit Tool
- Refactored applyEdit.ts from 1351 lines to modular structure (<300 lines per file)
- Fixed transaction result calculation bug to count actual edits
- Added comprehensive path traversal protection
- Moved dynamic imports to module level for performance
- Extracted schema definitions to separate file
- Improved code action selection logic
- Added MCPError class to common-types
- Updated error codes to include InternalError and InvalidRequest
