+++ src/tools/applyEdit/executors/index.ts
@@ -1,0 +1,8 @@
+export { executeCodeActions } from './codeActions.js';
+export { executeRename } from './rename.js';
+export { executeFormat, executeOrganizeImports } from './format.js';
+export { executeTextEdit, executeMultiFileEdit } from './textEdit.js';
+export { executeSearchReplace } from './searchReplace.js';
+export { executeFileOperations } from './fileOperations.js';
+export { executeSmartInsert } from './smartInsert.js';
+export { executeBatchOperations } from './batch.js';
